separate select token logic from bridge logic
